Backend/requirements.txt
flask
flask-cors
ultralytics
torch
torchvision
facenet-pytorch
opencv-python
numpy
Pillow
